move data migration code out of bin so it can be reused in background processor
